+++ MatrixKit/Utils/MXKConstants.m
@@ -17,7 +17,7 @@
 
 #import "MXKConstants.h"
 
-NSString *const MatrixKitVersion = @"0.9.1";
+NSString *const MatrixKitVersion = @"0.9.2";
 
 NSString *const kMXKErrorNotification = @"kMXKErrorNotification";